moving code into its own popnet module
