+++ bmtk/simulator/popnet/ssn/default_setters/__init__.py
@@ -1,0 +1,1 @@
+from . import activation_functions
