+++ bmtk/simulator/popnet/ssn/modules/__init__.py
@@ -1,0 +1,3 @@
+from .initial_states import InitStatesMod
+from .rates_recorder import RatesRecorderMod
+from .external_inputs import ExternalRatesMod
