+++ bmtk/simulator/popnet/ssn/utils.py
@@ -1,0 +1,8 @@
+from functools import wraps
+
+
+def empty_decorator(func):
+    @wraps(func)
+    def defunct(*args, **kwargs):
+        return func(*args, **kwargs)
+    return defunct
