+++ examples/ssn_l23/config.simulation.json
@@ -27,7 +27,8 @@
         "file": "$BASE_DIR/input/l4e_rates.h5",
         "node_set": {
             "population": "l4e"
-        }
+        },
+        "enabled": true
       },
       "bkg_inputs": {
         "input_type": "external_rates",
@@ -35,7 +36,8 @@
         "file": "$BASE_DIR/input/bkg_rates.npy",
         "node_set": {
             "population": "bkg"
-        }
+        },
+        "enabled": true
       }
     },
